Rename js file
